Make things more stable
